+++ src/config_game.h
@@ -49,9 +49,9 @@ class CGameConfig
 {
 public:
   uint32_t                 m_VoteKickPercentage;         // percentage of players required to vote yes for a votekick to pass
-  uint32_t                 m_NumPlayersToStartGameOver;  // when this player count is reached, the game over timer will start
-  uint32_t                 m_MaxPlayersLoopback;
-  uint32_t                 m_MaxPlayersSameIP;
+  uint8_t                  m_NumPlayersToStartGameOver;  // when this player count is reached, the game over timer will start
+  uint8_t                  m_MaxPlayersLoopback;
+  uint8_t                  m_MaxPlayersSameIP;
   uint32_t                 m_SyncLimit;                  // the maximum number of packets a player can fall out of sync before starting the lag screen (by default)
   uint32_t                 m_SyncLimitSafe;              // the maximum number of packets a player can fall out of sync before starting the lag screen (by default)
   bool                     m_SyncNormalize;              // before 3-minute mark, try to keep players in the game
